Multiple Language Support
### Description of Proposal

This proposal adds a human language local to USD so that content may be localized or provide hints for assistive technologies.

[Link to Rendered Proposal](https://github.com/dgovil/USD-proposals/blob/language/proposals/language/README.md)

### Disclaimer

This proposal isn't representative of upcoming functionality in our products. This is a future looking proposal that seemed timely to augment the proposals by others here like Autodesk.

This builds on discussion in [Issue 49](https://github.com/PixarAnimationStudios/OpenUSD-proposals/issues/49)

### Supporting Materials

An example

```
def Foo(
    prepend apiSchemas = ["LocaleAPI"]
    language = "en_US"
) {
    string text = "There's a snake in my boot"
    string text:fr_CA = "Il y a un serpent dans ma botte"
    string text:hi = "मेरे जूते में एक सांप है"
}
```
### Contributing

- [X] I agree to and accept the [Supplemental Terms](https://graphics.pixar.com/usd/release/contributing_supplemental.html).