Update the proposal according to the comments. (#17)
Update the proposal according to the comments.

We will focus on the schema in the proposal, not the implementation. So I removed the part about the Rprims, and also the plugins.
Add more detail about the SimpleText and MarkupText schemas. Add how the extents should be implemented.
Add the language tag for the textData in SimpleText and markup in MarkupText, and refer to the language proposal.
Add the "plain" tag for markup which is an alternate string when the markup can not be identified.
Remove the renderer property in SimpleText and MarkupText.
Change the name of the properties in TextStyle. Add altTypeface property which is used when the main typeface can not support some characters.
Add details about the font weight.
The TextLayout and TextLayoutAPI are merged into one schema. The new TextLayoutAPI not only contains the character directions, but also contains the direction how the lines are stacked.
Modify the name of the properties in ColumnStyle. Remove the direction and linesFlowDirection property in ColumnStyle. Instead, you can apply a TextLayoutAPI to a ColumnStyle to specify the directions.
In the part of "Some implementation details", discuss the extrusion, the missing characters, the layout for complex script and the billboard.
Finally there is a big part of four examples.

<!--
Put screenshots or other summary information here
-->

<!--
Please review the  [Contributing](https://graphics.pixar.com/usd/release/contributing_to_usd.html) page in the
documentation for the Supplemental Terms that apply to this repository.
Place an X in the box when you have reviewed and agree to the Supplemental Terms.
-->
- [ ] I agree to and accept the [Supplemental Terms](https://graphics.pixar.com/usd/release/contributing_supplemental.html).